requirements.txt
requests
feedparser
python-telegram-bot==20.8
schedule
beautifulsoup4
google-generativeai
cryptography